Mobile, Backend - nutrition api and barcode scanning added. Friends features added. Profile info added

+++ Mobile/DietApp/app/src/main/java/com/example/dietapp/models/dto/MealEntryDTO.kt
@@ -10,7 +10,7 @@ data class MealEntryDTO(
     val date: Date,
     val nutrients: Nutrients,
     val weightGram: Int,
-    val kcal: Int
+    val kcal: Double
 ) {
     fun toMealEntry() = MealEntry(id, name, date, nutrients, weightGram, kcal)
 }
